backend/requirements.txt
django==3.2
djangorestframework==3.12.4
djoser==2.1.0
drf-extra-fields==3.4.1
Pillow==9.4.0
pdfkit
django-colorfield
pytz
python-dotenv
psycopg2-binary==2.8.6
requests==2.26.0
PyJWT==2.4.0
pytest==6.2.4
pytest-pythonpath==0.7.3
django-filter==21.1
djangorestframework-simplejwt==4.7.2